Create image via setup script.
