+++ secrets/host_key
@@ -1,0 +1,1 @@
+ecdsa-sha2-nistp256 AAAAE2VjZHNhLXNoYTItbmlzdHAyNTYAAAAIbmlzdHAyNTYAAABBBBbo0vCu/TxD9NeYgxXliic7KuhrNpTIF+vKQ7hBo+DOqiyJHXeExf5KYMYX/wcc0vvS36P9IYkn85+dyh0kW90= root@wpia-didess1
