+++ secrets/ssh_key
@@ -1,0 +1,27 @@
+-----BEGIN RSA PRIVATE KEY-----
+MIIEogIBAAKCAQEAuDGfzrdeUFsZrtJwDdphBybJGp2a1n9CAUkAAzGKtxqBct4z
++q9IShh+NLxdRAcn1TuGooKX7Yl5KQawIXHOUMVcbWHwEqdUpUA4Eim6lrK6SvQr
+dCChGYGzkwklchpDDZHq/SRGWxLQU0HpfPl4QWlM4zVfmVZfMID+SBNs6EBXLv1q
+jcUrWGTmnc9CYHe3Oclag9Tba6YjZ9lwvxH7DPDCoMFU2kX8QvLV1+/PIoqPBwBq
+9A7qd7amxJqBCLLqUPkjsyctF0JALIZb61Miz0DBdhlv0KyVB4OIOn0g4BKy7u3Y
+qwMapHKJ9nb/C5ACNytT7zZnmSvBLq0hc+1uXQIDAQABAoIBACDx702ACsSApuVY
+CmbP6o7m5w11px46gMnODsyDbfY2LXUoPU/XTizCmMBB+MSUKxG+peQnS5lW/27R
+HWu9+VMFGfUOkWwktcUg6WaGRDN0IFILvkH1HBt90Wq5NhNkkx24PbmtZVAg21Yo
+8eRCkrnQZco137WDd18rk0+8+bvIiHx60n/8vSWcP2C+9JN8A7ai1SE9enCho/hk
+DuDp/0moFL0rbXuP6zWU4cyGvMXuFHxHHZC1J2C+clIr5PSTRPcyzixuI7NMzvro
+t9SaBdsKV6llHGefFBAYe3p0LLOtamaOojm+SoQeenSxYXWWswKRzvW1GP7sBCvv
+6g+BtZkCgYEA5swIdiQ/77Guo8Blm119qbSCVjuF9BABFrq+NTXmEyNWOeC0s5Fs
+2C/wu+XtCjxqD0hgqrfE9f0lVXUctQUI6mLgv06ouYhgITEx73h8nnVyjpmj/yeY
+SdrLL/MEkE3we0FXXpZvrwiXzzw0ekXYVga2rsWA159dyaeLKCslYvsCgYEAzE7K
+3vLGa5pG2/run+0Uttkg2MbrR6qzfnVI2VNAx6RBgjb7fmhPUXeZEWW0Ho5yCax2
+ZAiwVe2IjwPKsLR4TbYauIi9hTdtfkA3qr/CU9gpxvsarjADe/tOyHgMkOxqJ4SB
+4x56x6ya8jeqHu+7lzIkJ+dF3Gmv2TdJVARt9IcCgYAKsffOab2gwO4j3n2W5o01
+OQ01K8r4VWmHgR3hX18LEf8j7XrSO3H8zrPsMSp75/LORH9eSefRsmpxfFzZMqOI
+IQ7jLqT4/A3gW+RBgabcvPqTCyGjaFVRTtdNoIhcm5ACVjPTSUthpBdM3jBcXjI5
+QlRi4vomg9cXyjDTqrQv0wKBgCc1Q8ePbba+r7SNrnewhtYkSyJu4j4pqzKsjWNv
+j7IAerXjmVYD7Jzkx9JH404t5gKnPz0pUbXb1OYlQyOwLpozRWN+kmhbX77n0gcJ
+r+BpTnMqQ0+EVVbYvqDIMRkq1aBp1IdilQt2u3GabyJqvWRhsgCaaJTJD38vFfpy
+sytHAoGAFzW8+b/Ru0NgHBkrr/oFn7vKfoq1SFmtCiPBZAdHWzdazQo6BrhrZMKA
+QQkyDhQLYAj2AG0I5Q03BNiBT5SXmWB+qS+i3CTrj7WGhtPkUuYot36xOs70W53X
+OqooWA2TyrR07l1hzLMA7kjtNnJHLEIiYAvpvp5JXRVpwF0BdBQ=
+-----END RSA PRIVATE KEY-----
