+++ setup.sh
@@ -1,36 +1,27 @@
 #!/usr/bin/env bash
 set -e
 
-HERE=${BASH_SOURCE%/*}
-source ${HERE}/vault_auth.sh
-
-echo "Installing Pip and required Python packages"
-apt-get install python3-pip -y > /dev/null
-pip3 install -q -r ${HERE}/requirements.txt
-
-echo "-------------------------------------------"
-echo "Installing bb8 user and group"
-
-if ! id -u bb8 > /dev/null 2>&1; then
-    useradd bb8 -U -d /var/lib/bb8
-    password=$(vault read --field password secret/backup/bb8/user)
-    echo "bb8:$password" | chpasswd
-
-    mkdir -p /var/lib/bb8/.ssh
-    
-    # add to docker group
-    usermod -aG docker bb8
+if [[ ($# -eq 0) ]]; then
+    echo "Usage: ./setup.sh PATH_TO_SOURCE_CONFIG [TARGET ...]"
+    exit -1;
 fi
 
-# give bb8 group ownership of this dir
-chgrp -R bb8 $HERE
-chmod -R g+w $HERE
+source_config_path=$1
+shift && targets="$@"   # Targets is all the args after the first
 
-echo "-------------------------------------------"
-echo "Running setup.py:"
-(cd "$HERE" && su -c ./setup.py bb8)
+# Get secrets and put them in secrets/
+# (where the Dockerfile looks for them)
+HERE=${BASH_SOURCE%/*}
+source ${HERE}/vault_auth.sh
+./obtain_secrets.py
 
-echo "-------------------------------------------"
+# Copy config
+cp "$source_config_path" source-config.json
 
-echo "Setup complete. To schedule backups, run sudo ./schedule.py"
-echo "To perform a restore, run ./restore.py"
+docker build --build-arg "TARGETS=$targets" --tag bb8 .
+
+echo "-----------------------------------------------"
+echo "Setup complete. You can now: "
+echo "backup:           Run ./backup.sh"
+echo "schedule backups: Run sudo ./schedule.sh"
+echo "restore:          Run ./restore.sh"
